Credit Card Validator and Analyzer - Azure AI/src/requirements.txt
azure.core
azure-ai-documentintelligence
streamlit
azure-storage-blob
python-dotenv

#Use pip install -r requirements.txtS